Don't forget basic attacks too
