+++ module/checks/check-configuration.mjs
@@ -8,7 +8,7 @@ import { PlayerListEnhancements } from '../helpers/player-list-enhancements.mjs'
 import { Targeting } from '../helpers/targeting.mjs';
 import { DamageData } from './damage-data.mjs';
 import { UpdateResourceData } from '../pipelines/resource-pipeline.mjs';
-import { ChooseWeaponDialog } from '../documents/items/skill/choose-weapon-dialog.mjs';
+import { WeaponResolver } from '../documents/items/skill/weapon-resolver.mjs';
 import FoundryUtils from '../helpers/foundry-utils.mjs';
 
 // Data keys
@@ -407,7 +407,7 @@ export class CheckConfigurer extends CheckInspector {
 	 * @returns {CheckConfigurer}
 	 */
 	addWeaponAccuracy(weapon) {
-		const baseAccuracy = ChooseWeaponDialog.getAccuracy(weapon);
+		const baseAccuracy = WeaponResolver.getAccuracy(weapon);
 		if (baseAccuracy) {
 			this.addModifier('FU.AccuracyCheckBaseAccuracy', baseAccuracy);
 		}
